Add extension skeleton, simple communication
There is the extension skeleton added to the project. It contains
the necessary manifest + logo. On top of that, there are several
scripts.

Background represents the background page, which is able to mediate
the communication between Popup and Content Script. The content script
itself uses another JS file (Injected), in order to be able to access
the page variables + DOM.